Refactor audio
I moved the native audio implementation into crates/audio/src/native.rs
and then changed the audio crate so that it exports the native audio
implementation in native builds or the audio wrapper in web builds.
